Add thumbnail/full_size_picture fields to FormUrl
Add optional thumbnail and full_size_picture fields to FormUrl.
If they are specified, the thumbnail url will be rendered in the component
and the full_size_picture passed on to the image viewer when the user touches a picture.
This makes it possible to display thumbnails in the form and only load the full-size picture
when the users opens the image viewer ; thus making the initial load faster.
It also makes it possible to display smaller pictures in the form body,
to make it more readable on smaller phones